+++ synchronization/rcu/info.sh
@@ -1,13 +1,23 @@
 #!/bin/bash
 set -e
 
-print_param() {
+rcutree_param() {
 	local param=$1
 	local fmt=%lx
 	[[ $2 ]] && fmt=$2
 	echo ${param} = $(printf "${fmt}\n" $(cat /sys/module/rcutree/parameters/${param}))
 }
 
-print_param jiffies_till_sched_qs
-print_param blimit %ld
-print_param gp_cleanup_delay %ld
+rcupdate_param() {
+	local param=$1
+	local fmt=%lx
+	[[ $2 ]] && fmt=$2
+	echo ${param} = $(printf "${fmt}\n" $(cat /sys/module/rcupdate/parameters/${param}))
+}
+
+rcutree_param jiffies_till_sched_qs
+rcutree_param blimit %ld
+rcutree_param gp_cleanup_delay %ld
+
+rcupdate_param rcu_cpu_stall_timeout %ld
+rcupdate_param rcu_exp_cpu_stall_timeout %ld